requirements.txt
Django==2.1.1
djangorestframework==3.9
django-filter
django-crispy-forms
django-cors-headers
channels==1.1.8
channels_api
asgi_redis
psycopg2-binary
gunicorn
django-heroku
markdown
drf-yasg
flex
swagger_spec_validator
